+++ src/ngscopeclient/MetricsDialog.h
@@ -47,6 +47,15 @@ public:
 
 protected:
 
+	std::string FormatValueWithPercentage(int64_t a, int64_t total)
+	{
+		float percent = (float)a * 100 / (float)total;
+
+		char tmp[128];
+		snprintf(tmp, sizeof(tmp), "%" PRIi64 " (%.2f %%)", a, percent);
+		return tmp;
+	}
+
 	int m_displayRefreshRate;
 };